+++ src/test/kotlin/blackjack/core/CardDispenserTest.kt
@@ -1,6 +1,7 @@
 package blackjack.core
 
 import blackjack.core.card.Card
+import blackjack.core.card.CardDispenser
 import blackjack.core.player.Name
 import blackjack.core.player.Player
 import blackjack.core.player.Players
