+++ src/test/kotlin/blackjack/core/turn/TurnTest.kt
@@ -1,6 +1,6 @@
 package blackjack.core.turn
 
-import blackjack.core.CardDispenser
+import blackjack.core.card.CardDispenser
 import blackjack.core.card.Cards
 import blackjack.core.player.Name
 import blackjack.core.player.Player
